pyproject.toml
[project]
name = "multimodal-multi-agent-image-workflow"
version = "0.1.0"
description = "A modular multimodal multi-agent pipeline for image understanding, prompt refinement, img2img generation, and evaluation."
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "openai>=1.30",
  "replicate>=0.25",
  "requests>=2.32",
  "python-dotenv>=1.0",
  "Pillow>=10.0",
  "torch>=2.2",
  "transformers>=4.41",
]

[tool.ruff]
line-length = 100
target-version = "py310"
